CLI-1078 Move CliCommandExecuted into its own telemetry module
