+++ src/core/telemetry/command-executed.ts
@@ -1,0 +1,56 @@
+/*
+ * SonarQube CLI
+ * Copyright (C) SonarSource Sàrl
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { TelemetryFact } from '@/core/commands/invocation-context.ts';
+import type { SonarCommand } from '@/core/commands/sonar-command.ts';
+import { DISTRIBUTION, type Distribution } from '@/core/host/distribution.ts';
+
+import { currentProjectUuid } from './project-uuid.ts';
+
+export const CLI_COMMAND_EXECUTED = 'CliCommandExecuted';
+
+/** Domain payload for CliCommandExecuted (identity is filled at drain time). */
+export type CommandExecutedPayload = {
+  command: string | undefined;
+  subcommand: string | null;
+  result: 'success' | 'failure';
+  distribution: Distribution;
+  project_uuid: string | null;
+};
+
+/**
+ * Build a CliCommandExecuted fact for a finished command.
+ *
+ * `result` is derived from `process.exitCode` (`success` when 0 or unset).
+ * `project_uuid` is resolved here (async, never rejects). Identity is applied at commit.
+ * Command/subcommand come from {@link SonarCommand.commandAndSubcommand}.
+ */
+export async function buildCommandExecutedFact(
+  command: SonarCommand,
+): Promise<TelemetryFact<CommandExecutedPayload>> {
+  const { command: commandName, subcommand } = command.commandAndSubcommand();
+  return new TelemetryFact(CLI_COMMAND_EXECUTED, {
+    command: commandName,
+    subcommand,
+    result: (process.exitCode ?? 0) === 0 ? 'success' : 'failure',
+    distribution: DISTRIBUTION,
+    project_uuid: await currentProjectUuid(),
+  });
+}
